refactor: extract components, fix style
