+++ src/components/gallery/Gallery.scss
@@ -23,7 +23,7 @@
   }
 }
 
-.wrapper__section {
+.section__wrapper {
   height: 100%;
   padding: 84px 135px 0;
   display: flex;
