+++ src/styles/_variables.scss
@@ -1,6 +1,7 @@
 $color-white: #ffffff;
 $color-white-opacity-29: rgb(255, 255, 255, .29);
 $color-orange: #EF6D58;
+$color-orange-dark: #e44128;
 $color-brown: #391400;
 $color-primary: #FDF0E9;
 $color-primary-dark: #F3D1BF;
