+++ src/styles/index.scss
@@ -30,7 +30,7 @@ body {
     background:  $color-orange;
 
     &:hover {
-      background: $color-primary-dark;
+      background: $color-orange-dark;
     }
   }
 }
